Exercise 4
Got 3 of the four test to pass.

I have a sneaky suspicion that test one was not supposed to have a
period at the end of the sentence. But I could be wrong.

Still need to figure out how to make the empty string test pass. I think
it has to do with the Array indexing. But not sure what just yet.

Arrays index from 0, correct? And the test is looking for that, but I
think that when the method runs for the testEmptyString test it is
running through, getting an array index of 1 and test is failing. The
array length is 1 where the test is looking for 0. 

I sorta cheated and added quotes to: expectedWords = new String[] {""};
like so and the test passed. 

I took them out and I'm trying to figure out my code to see if I can
change something there, but have not got it yet.

Or maybe it is counting that whitespace, which i need to figure out how
to remove that like i did with the Regex //W, but I cannot figure just
how to do that.

Also, without Internet access, i have no access to JavaDocs....which
really sucks.