This is where I stop for awhile.
So as we talked about on FB. I got two of the test to pass. But one of
them is failing due to the word "don't". That appostrophe is fucking me
up.

I about 95% sure it is due to my not getting my Regex syntax correct.

But i'm gonna commit what I got and come back to it when I get back to
aplace I can work on this some more. Probably won't be for a few weeks.